Refactor DTOs and models to use BigDecimal for monetary values; update PedidoServiceImpl for improved data handling and validation

+++ BackTienda/src/main/java/com/example/BackTienda/dto/PedidoItemDTOs/PedidoItemResponseDTO.java
@@ -1,5 +1,7 @@
 package com.example.BackTienda.dto.PedidoItemDTOs;
 
+import java.math.BigDecimal;
+
 import lombok.Data;
 
 @Data
@@ -9,6 +11,6 @@ public class PedidoItemResponseDTO {
     private Long productoId;
     private String nombreProducto;
     private int cantidad;
-    private double precioUnitario;
-    private double subtotal;
+    private BigDecimal precioUnitario;
+    private BigDecimal subtotal;
 }
